Make compiler happy about size_t -> int conversion.
PiperOrigin-RevId: 321060817

+++ common/formatting/align_test.cc
@@ -369,8 +369,8 @@ TEST_F(Sparse3x3MatrixAlignmentMoreSpacesTest,
   TabularAlignTokens(
       &partition_, kDefaultAlignmentHandler, &pre_format_tokens_, sample_,
       // Alignment disabled over line 2
-      ByteOffsetSet(
-          {{sample_.find_first_of("\n") + 1, sample_.find("four") + 4}}),
+      ByteOffsetSet({{static_cast<int>(sample_.find_first_of("\n") + 1),
+                      static_cast<int>(sample_.find("four") + 4)}}),
       40);
 
   EXPECT_EQ(pre_format_tokens_[1].before.break_decision,
